+++ android/app/src/main/java/com/on/turip/domain/videoinfo/TravelCourseRepository.kt
@@ -1,0 +1,4 @@
+package com.on.turip.domain.videoinfo
+
+interface TravelCourseRepository {
+}
