+++ android/app/src/main/java/com/on/turip/domain/videoinfo/contents/Content.kt
@@ -1,4 +1,6 @@
-package com.on.turip.domain.contents
+package com.on.turip.domain.videoinfo.contents
+
+import com.on.turip.domain.videoinfo.contents.creator.Creator
 
 data class Content(
     val id: Long,
