+++ android/app/src/main/java/com/on/turip/domain/videoinfo/contents/PagedContentsResult.kt
@@ -1,4 +1,4 @@
-package com.on.turip.domain.contents
+package com.on.turip.domain.videoinfo.contents
 
 data class PagedContentsResult(
     val videos: List<VideoInformation>,
