+++ android/app/src/main/java/com/on/turip/domain/videoinfo/contents/VideoInformation.kt
@@ -1,4 +1,4 @@
-package com.on.turip.domain.contents
+package com.on.turip.domain.videoinfo.contents
 
 data class VideoInformation(
     val content: Content,
