+++ android/app/src/main/java/com/on/turip/domain/videoinfo/contents/creator/repository/CreatorRepository.kt
@@ -1,0 +1,4 @@
+package com.on.turip.domain.videoinfo.contents.creator.repository
+
+interface CreatorRepository {
+}
